refactor: fix type in exception package
